New module for the builder
`ir::Builder` was getting pretty big, so it is now `builder::Builder`
instead.
