+++ flatgfa-sh/src/main.rs
@@ -1,3 +1,4 @@
+mod builder;
 mod eval;
 mod ir;
 mod opt;
